basic support for terminal resizing
Not realtime, the size change is detected and used on the next run.

This definitely doesn't handle everything, but it at
least makes things print correctly on the next run.

Though there can still be some garbage-looking output above the printed
output when it doesn't take up the full screen.