Fix display for longer command lines
This largely means being more careful about how we move the cursor
around, to avoid doing the wrong thing when the command line wraps
or is exactly one screen width.

We now "reset" the cursor to the beginning of the command line before
doing anything else, it's just too hard to move the cursor consistently
otherwise. The readline redraw takes care of moving the cursor back
to its current place in the command.

This fixes #10 and some other issues that I found while testing
the change.